requirements.txt
# --- Core Framework ---
fastapi
uvicorn

# --- LangChain & Embeddings ---
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
tiktoken

# --- Utilities ---
python-dotenv
openai

# --- Document Loading / Parsing ---
unstructured
unstructured[local-inference]
